Reorganize file structure to avoid circular imports
