--- browser_use/agent/prompts/__init__.py
@@ -1,7 +1,0 @@
-"""
-Prompts module for browser-use agent.
-"""
-
-from ..prompts import SystemPrompt, AgentMessagePrompt, PlannerPrompt
-
-__all__ = ['SystemPrompt', 'AgentMessagePrompt', 'PlannerPrompt']
